refactor(background-activity): keep business services platform-blind
KeepAliveCoordinator now selects one registered lease source at
construction (AudioKeepAliveSource on iOS, AndroidBackgroundActivityRuntime
on Android, no-op elsewhere) instead of inlining audio and branching on
Platform.OS. The iOS audio mechanism moves unchanged into
AudioKeepAliveSource. BackgroundReplyRuntime and BackgroundActivityManager
drop their platform whitelists and rely on no-op presenters and sources.
The task deep-link URL contract lives once in shared/backgroundActivity;
backend builds and App Shell parses through it. Docs describe the source
model and the Android foreground service.
