+++ src/backend/core/application/__tests__/serviceRegistry.test.ts
@@ -128,7 +128,11 @@ describe('service registry', () => {
     expect(getAppStatePolicy(services.JobRuntime)).toBe('continue');
     expect(getAppStatePolicy(services.BackgroundReplyRuntime)).toBe('background-presentation');
     expect(getAppStatePolicy(services.BackgroundActivityManager)).toBe('background-presentation');
-    expect(getAppStatePolicy(services.KeepAliveCoordinator)).toBe('background-presentation');
+    expect(getAppStatePolicy(services.KeepAliveCoordinator)).toBe('not-applicable');
+    expect(getAppStatePolicy(services.AudioKeepAliveSource)).toBe('background-presentation');
+    expect(getAppStatePolicy(services.AndroidBackgroundActivityRuntime)).toBe(
+      'background-presentation',
+    );
     expect(getAppStatePolicy(services.ProviderRegistryUpdaterService)).toBe('not-applicable');
   });
