+++ src/frontend/appShell/backgroundActivity/__tests__/backgroundActivityNavigation.test.ts
@@ -1,22 +1,7 @@
 import { backgroundActivityHref } from '../backgroundActivityNavigation';
 
-// Expo Linking reads native execution metadata even when parsing an explicit
-// URL. Give its real parser the custom development client's environment.
-jest.mock('expo-constants', () => {
-  const actual = jest.requireActual<typeof import('expo-constants')>('expo-constants');
-  return {
-    ...actual,
-    __esModule: true,
-    default: {
-      ...actual.default,
-      executionEnvironment: actual.ExecutionEnvironment.Bare,
-      expoConfig: { scheme: ['cherrystudio', 'cherrystudio-dev', 'cherrystudio-preview'] },
-    },
-  };
-});
-
 test.each(['cherrystudio', 'cherrystudio-dev', 'cherrystudio-preview'])(
-  'resolves task links for the %s variant without opening arbitrary destinations',
+  'routes task links for the %s variant and opens nothing for other destinations',
   (scheme) => {
     expect(backgroundActivityHref(`${scheme}:///?agentId=a&sessionId=s`, scheme)).toEqual({
       pathname: '/',
